// generated/src/app/api-models/stage-claim-oc.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageClaimOc {
  batchId: string;
  transactionId: number;
  sequenceNo: number;
  claimFilingInd: string;
  sourceOfPay: string;
  insuranceTypeCode: string;
  payorName: string;
  payorAddress1: string;
  payorAddress2: string;
  payorCity: string;
  payorState: string;
  payorPostalCode: string;
  payorCountry: string;
  payorOrganizationId: string;
  groupName: string;
  groupNo: string;
  assignOfBenefits: string;
  patSignatureSource: string;
  patRelToInsured: string;
  patReleaseInd: string;
  insuredIdNo: string;
  insuredLastName: string;
  insuredFirstName: string;
  insuredMiddleName: string;
  insuredGeneration: string;
  insuredGender: string;
  insuredDob: Date;
  insuredEmplymentStatus: string;
  insuredEmployerLoc: string;
  insuredEmployerName: string;
  disallowedCostCont: number;
  disallowedOther: number;
  allowedAmt: number;
  deductibleAmt: number;
  coinsuranceAmt: number;
  priorPayment: number;
  adjudInd1: string;
  adjudInd2: string;
  adjudInd3: string;
  balanceDue: number;
  internalControlNo: string;
  providerNo: string;
  treatmentAuthCode: string;
  refProviderNo: string;
  insCardEffDate: Date;
  insCardTermDate: Date;
  insLocationId: string;
  priorAuthNo: string;
  payorClaimOfficeNo: string;
  planType: string;
  planPercentage: number;
  policyNo: string;
  insuredEmployerCity: string;
  insuredEmployerState: string;
  insuredEmployerPostalCode: string;
  insertDatetime: Date;
}